refactor: update line drawing algorithm to support float coordinates

+++ include/line.h
@@ -3,9 +3,4 @@
 #include <functional>
 
 // modified Gupta-Sproull algorithm
-void line(int32_t x1,
-          int32_t x2,
-          int32_t y1,
-          int32_t y2,
-          double t,
-          const std::function<void(int32_t, int32_t, double)>& f);
+void line(double x1, double x2, double y1, double y2, double t, const std::function<void(int32_t, int32_t, double)>& f);
